change token id from u32 to u64 (#1816)
* change token id from u32 to u64

This is done because it is conceivable that we will have more than
4 billion tokens on our chain, particularly if this mechanism is
used to create NFTs or something someday.

There is very little reason from an engineering point of view to
try to save space by using a u32 instead of a u64 -- u32's do
sometimes run out in practice, and a u64 will not here.

* Update consensus/mint-client/src/config.rs


* Update consensus/mint-client/src/config.rs


* Update transaction/core/src/amount/mod.rs


* fix build

+++ mint-auditor/src/db.rs
@@ -51,8 +51,8 @@ pub const COUNTERS_KEY: &str = "counters";
 #[derive(Deserialize, Eq, Message, PartialEq, Serialize)]
 pub struct BlockAuditData {
     /// A map of token id -> calculated balance.
-    #[prost(btree_map = "uint32, uint64", tag = 1)]
-    pub balance_map: BTreeMap<u32, u64>,
+    #[prost(btree_map = "uint64, uint64", tag = 1)]
+    pub balance_map: BTreeMap<u64, u64>,
 }
 
 /// Statistics we keep track of.
